Refactor and take extensions into consideration in dynflags
