+++ ChangeLog.md
@@ -1,4 +1,11 @@
 # Changelog for ghc-lib-parser-ex
 
+## 8.8.3.0 released 2020-01-25
+- Change in versioning scheme;
+- New modules:
+  - `Language.Haskell.GhclibParserEx.Config`
+  - `Language.Haskell.GhclibParserEx.DynFlags`
+- `parsePragmasIntoDynFlags` signature change.
+
 ## 8.8.1.20191204, 8.8.2, 0.20200102 released 2020-01-18
 - First releases
