+++ azure-pipelines.yml
@@ -25,32 +25,32 @@ strategy:
       image: "macOS-10.14"
       stack: "stack.yaml"
 
-    linux-881-881:
+    linux-881-88:
       image: "Ubuntu 16.04"
       stack: "stack-808-808-ghc-lib.yaml"
-    windows-881-881:
+    windows-881-88:
       image: "vs2017-win2016"
       stack: "stack-808-808-ghc-lib.yaml"
-    mac-881-881:
+    mac-881-88:
       image: "macOS-10.14"
       stack: "stack-808-808-ghc-lib.yaml"
     linux-881-881-da:
       image: "Ubuntu 16.04"
       stack: "stack-da-8.8.1.20200122.yaml"
-    windows-881-881-da:
+    windows-881-88-da:
       image: "vs2017-win2016"
       stack: "stack-da-8.8.1.20200122.yaml"
-    mac-881-881-da:
+    mac-881-881-881-da:
       image: "macOS-10.14"
       stack: "stack-da-8.8.1.20200122.yaml"
 
-    linux-881-881-no-ghc-lib:
+    linux-881-88-no-ghc-lib:
       image: "Ubuntu 16.04"
       stack: "stack-808-808-no-ghc-lib.yaml"
-    windows-881-881-no-ghc-lib:
+    windows-881-88-no-ghc-lib:
       image: "vs2017-win2016"
       stack: "stack-808-808-no-ghc-lib.yaml"
-    mac-881-881-no-ghc-lib:
+    mac-881-88-no-ghc-lib:
       image: "macOS-10.14"
       stack: "stack-808-808-no-ghc-lib.yaml"
