+++ stack-808-808-ghc-lib.yaml
@@ -1,5 +1,5 @@
 resolver: nightly-2020-01-08 # ghc-8.8.1
-extra-deps: [ghc-lib-parser-8.8.1.20191204] # ghc-lib-parser-8.8.1
+extra-deps: [ghc-lib-parser-8.8.2]
 ghc-options:
     "$locals": -Wall -Wno-name-shadowing
 flags:
